Refactored anatomy. Additionally, moved all common functions to common_utils. FUTURE: put functions in classes
